==> notebooks/02_feature_engineering.ipynb <==
# CASE: BOMBOM - Advanced Analytics IBM + Vale

## Realizado por: Sabrina Otoni da Silva

## Etapa de feature engineering

from pathlib import Path

import pandas as pd

from sklearn.preprocessing import OneHotEncoder, OrdinalEncoder
from sklearn.pipeline import Pipeline
from sklearn.compose import ColumnTransformer

import joblib

datapath = Path('../data')

data_model = pd.read_excel(f'{datapath}/01_raw/registros-prod.xlsx')
data_predictive = pd.read_excel(f'{datapath}/01_raw/analise-preditiva.xlsx')
data_prescriptive = pd.read_excel(f'{datapath}/01_raw/analise-prescritiva.xlsx')

categorical_hot_transformer = Pipeline(steps=[
    ('onehot', OneHotEncoder(handle_unknown='ignore', sparse_output=False, drop='first'))
])

categorical_ordinal_transformer = Pipeline(steps=[
    ('ordinal', OrdinalEncoder(handle_unknown='error'))
])

preprocessor_hot = ColumnTransformer(
    transformers=[
        ('cat', categorical_hot_transformer, ['VAR_2'])
    ])

preprocessor_ordinal = ColumnTransformer(
    transformers=[
        ('cat', categorical_ordinal_transformer, ['VAR_2'])
    ])

preprocessor_hot.fit(data_model)
preprocessor_ordinal.fit(data_model)

preprocessorpath = Path('../preprocessing')

joblib.dump(preprocessor_hot, preprocessorpath / 'preprocessor_hot.pkl')
joblib.dump(preprocessor_ordinal, preprocessorpath / 'preprocessor_ordinal.pkl')

data_model_hot_encoded = preprocessor_hot.transform(data_model)
data_predictive_hot_encoded = preprocessor_hot.transform(data_predictive)
data_prescriptive_hot_encoded = preprocessor_hot.transform(data_prescriptive)

data_model_ordinal_encoded = preprocessor_ordinal.transform(data_model)
data_predictive_ordinal_encoded = preprocessor_ordinal.transform(data_predictive)
data_prescriptive_ordinal_encoded = preprocessor_ordinal.transform(data_prescriptive)

data_model_hot_encoded = pd.DataFrame(data_model_hot_encoded, columns=preprocessor_hot.get_feature_names_out())
data_model_hot_encoded.columns = [col.split("__")[-1] for col in data_model_hot_encoded.columns]
data_model_hot_encoded = pd.concat([data_model, data_model_hot_encoded], axis=1)
data_model_hot_encoded.drop('VAR_2', axis=1, inplace=True)
data_model_hot_encoded.head()

data_predictive_hot_encoded = pd.DataFrame(data_predictive_hot_encoded, columns=preprocessor_hot.get_feature_names_out())
data_predictive_hot_encoded.columns = [col.split("__")[-1] for col in data_predictive_hot_encoded.columns]
data_predictive_hot_encoded = pd.concat([data_predictive, data_predictive_hot_encoded], axis=1)
data_predictive_hot_encoded.drop('VAR_2', axis=1, inplace=True)
data_predictive_hot_encoded.head()

data_prescriptive_hot_encoded = pd.DataFrame(data_prescriptive_hot_encoded, columns=preprocessor_hot.get_feature_names_out())
data_prescriptive_hot_encoded.columns = [col.split("__")[-1] for col in data_prescriptive_hot_encoded.columns]
data_prescriptive_hot_encoded = pd.concat([data_prescriptive, data_prescriptive_hot_encoded], axis=1)
data_prescriptive_hot_encoded.drop('VAR_2', axis=1, inplace=True)
data_prescriptive_hot_encoded.head()

data_model_hot_encoded.to_csv(f'{datapath}/02_intermediate/data_model_hot_encoded.csv', index=False)
data_predictive_hot_encoded.to_csv(f'{datapath}/02_intermediate/data_predictive_hot_encoded.csv', index=False)
data_prescriptive_hot_encoded.to_csv(f'{datapath}/02_intermediate/data_prescriptive_hot_encoded.csv', index=False)

data_model_ordinal_encoded = pd.DataFrame(data_model_ordinal_encoded, columns=['VAR_2'])
data_model.drop('VAR_2', axis=1, inplace=True)
data_model = pd.concat([data_model, data_model_ordinal_encoded], axis=1)
data_model.head()

data_predictive_ordinal_encoded = pd.DataFrame(data_predictive_ordinal_encoded, columns=['VAR_2'])
data_predictive.drop('VAR_2', axis=1, inplace=True)
data_predictive = pd.concat([data_predictive, data_predictive_ordinal_encoded], axis=1)
data_predictive.head()

data_prescriptive_ordinal_encoded = pd.DataFrame(data_prescriptive_ordinal_encoded, columns=['VAR_2'])
data_prescriptive.drop('VAR_2', axis=1, inplace=True)
data_prescriptive = pd.concat([data_prescriptive, data_prescriptive_ordinal_encoded], axis=1)
data_prescriptive.head()

data_model.to_csv(f'{datapath}/02_intermediate/data_model_ordinal_encoded.csv', index=False)
data_predictive.to_csv(f'{datapath}/02_intermediate/data_predictive_ordinal_encoded.csv', index=False)
data_prescriptive.to_csv(f'{datapath}/02_intermediate/data_prescriptive_ordinal_encoded.csv', index=False)
