==> notebooks/05_final_process.ipynb <==
# CASE: BOMBOM - Advanced Analytics IBM + Vale

## Realizado por: Sabrina Otoni da Silva

## Etapa de treinamento, predição e otimização (redução do custo)

from pathlib import Path

import numpy as np
import pandas as pd

from sklearn.ensemble import RandomForestRegressor

from hyperopt import hp, fmin, tpe

import joblib

import warnings
warnings.simplefilter("ignore", UserWarning)

datapath = Path('../data')

data_model = pd.read_csv(f'{datapath}/02_intermediate/data_hot_processed.csv')
data_predictive = pd.read_csv(f'{datapath}/02_intermediate/data_predictive_hot_encoded.csv')
data_prescriptive = pd.read_csv(f'{datapath}/02_intermediate/data_prescriptive_hot_encoded.csv')

X_train = data_model.drop(['VAR_2_A', 'PESO_BOMBOM'], axis=1)
y_train = data_model['PESO_BOMBOM']

rf_model = RandomForestRegressor(max_depth = 10, n_estimators = 50)
rf_model.fit(X_train.values, y_train.values)
modelpath = Path('../model')
joblib.dump(rf_model, modelpath / 'rf_model.pkl')

rf_model = joblib.load(modelpath / 'rf_model.pkl')

data_predictive['PESO_BOMBOM'] = rf_model.predict(data_predictive[['QTD_CHOC', 'VAR_1', 'VAR_2_B', 'VAR_2_C']])

def cost_function(p):
    return 200 / (1 + np.exp(10 * (p - 9.5))) + 200 / (1 + np.exp(-0.8 * (p - 12)))

def optimize_qtd_choc(row):
    space = hp.uniform('QTD_CHOC', data_model['QTD_CHOC'].min(), data_model['QTD_CHOC'].max())

    def objective(qtd_choc):
        predicted_weight = rf_model.predict(np.array([[qtd_choc, row['VAR_1'], row['VAR_2_B'], row['VAR_2_C']]]))[0]
        cost = cost_function(predicted_weight)
        return cost, predicted_weight

    def objective_wrapper(qtd_choc):
        return objective(qtd_choc)[0]

    best = fmin(
        fn=objective_wrapper,
        space=space,
        algo=tpe.suggest,
        max_evals=100,
        rstate=np.random.default_rng(0)
    )
    _, predicted_weight = objective(best['QTD_CHOC'])
    return best['QTD_CHOC'], predicted_weight

results = data_prescriptive.apply(lambda row: optimize_qtd_choc(row) if pd.isnull(row['QTD_CHOC']) else (row['QTD_CHOC'], np.nan), axis=1)
data_prescriptive['QTD_CHOC'] = results.apply(lambda x: x[0])
data_prescriptive['PESO_BOMBOM_PREV'] = results.apply(lambda x: x[1])

data_predictive.to_csv(f'{datapath}/03_processed/data_predictive_processed.csv', index=False)
data_prescriptive.to_csv(f'{datapath}/03_processed/data_prescriptive_processed.csv', index=False)

data_predictive.head(10)

data_predictive.describe()

data_prescriptive.head(10)

data_prescriptive.describe()
